pyproject.toml
[tool.poetry]
name = "preppin-data"
version = "0.1.0"
description = "My solutions to the Preppin' Data challenges"
authors = ["Vincent Ging Ho Yim <cenviity@gmail.com>"]
readme = "README.md"
packages = [{ include = "preppin_data", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
jupyter = "^1.0.0"
loguru = "^0.7.2"
pandas = { extras = ["excel"], version = "^2.2.0" }
pyarrow = "^15.0.0"
sf-hamilton = { extras = ["visualization"], version = "^1.67.0" }

[tool.poetry.group.dev.dependencies]
gdown = "^5.1.0"
rich = "^13.7.1"
ruff = "^0.4.8"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",
    "D212",
    "TCH003",
    # Rules that conflict with Ruff formatter
    "COM812",
    "COM819",
    "D206",
    "D300",
    "E111",
    "E114",
    "E117",
    "ISC001",
    "ISC002",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "W191",
]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
